+++ wizard/src/style.scss
@@ -147,6 +147,10 @@ h2 {
 
 /* Mapping Section */
 
+.mapping-section {
+  position: relative;
+}
+
 .highlighted-col {
   background-color: lightyellow;
 }
